moved utility functions & constants to detector.py
